PLFM-6426: Move the S3 bucket builder to its own main
