+++ src/main/java/org/sagebionetworks/template/repo/RepositoryTemplateBuilderImpl.java
@@ -1,12 +1,57 @@
 package org.sagebionetworks.template.repo;
 
+import static org.sagebionetworks.template.Constants.CAPABILITY_NAMED_IAM;
+import static org.sagebionetworks.template.Constants.CLOUDWATCH_LOGS_DESCRIPTORS;
+import static org.sagebionetworks.template.Constants.DATABASE_DESCRIPTORS;
+import static org.sagebionetworks.template.Constants.DB_ENDPOINT_SUFFIX;
+import static org.sagebionetworks.template.Constants.ENCRYPTED_AMI_IMAGE_ID;
+import static org.sagebionetworks.template.Constants.ENVIRONMENT;
+import static org.sagebionetworks.template.Constants.EXCEPTION_THROWER;
+import static org.sagebionetworks.template.Constants.INSTANCE;
+import static org.sagebionetworks.template.Constants.JSON_INDENT;
+import static org.sagebionetworks.template.Constants.OAUTH_ENDPOINT;
+import static org.sagebionetworks.template.Constants.OUTPUT_NAME_SUFFIX_REPOSITORY_DB_ENDPOINT;
+import static org.sagebionetworks.template.Constants.PARAMETER_MYSQL_PASSWORD;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_BEANSTALK_HEALTH_CHECK_URL;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_BEANSTALK_MAX_INSTANCES;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_BEANSTALK_MIN_INSTANCES;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_BEANSTALK_NUMBER;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_BEANSTALK_SSL_ARN;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_BEANSTALK_VERSION;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_INSTANCE;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_OAUTH_ENDPOINT;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_REPO_RDS_ALLOCATED_STORAGE;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_REPO_RDS_INSTANCE_CLASS;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_REPO_RDS_IOPS;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_REPO_RDS_MAX_ALLOCATED_STORAGE;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_REPO_RDS_MULTI_AZ;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_REPO_RDS_STORAGE_TYPE;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_ROUTE_53_HOSTED_ZONE;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_STACK;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_TABLES_INSTANCE_COUNT;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_TABLES_RDS_ALLOCATED_STORAGE;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_TABLES_RDS_INSTANCE_CLASS;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_TABLES_RDS_IOPS;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_TABLES_RDS_MAX_ALLOCATED_STORAGE;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_TABLES_RDS_STORAGE_TYPE;
+import static org.sagebionetworks.template.Constants.PROPERTY_KEY_VPC_SUBNET_COLOR;
+import static org.sagebionetworks.template.Constants.REPO_BEANSTALK_NUMBER;
+import static org.sagebionetworks.template.Constants.SHARED_EXPORT_PREFIX;
+import static org.sagebionetworks.template.Constants.SHARED_RESOUCES_STACK_NAME;
+import static org.sagebionetworks.template.Constants.SOLUTION_STACK_NAME;
+import static org.sagebionetworks.template.Constants.STACK;
+import static org.sagebionetworks.template.Constants.STACK_CMK_ALIAS;
+import static org.sagebionetworks.template.Constants.TEMPALTE_BEAN_STALK_ENVIRONMENT;
+import static org.sagebionetworks.template.Constants.TEMPALTE_SHARED_RESOUCES_MAIN_JSON_VTP;
+import static org.sagebionetworks.template.Constants.VPC_EXPORT_PREFIX;
+import static org.sagebionetworks.template.Constants.VPC_SUBNET_COLOR;
+
 import java.io.StringWriter;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
 import java.util.StringJoiner;
 
-import com.amazonaws.services.cloudformation.model.Tag;
 import org.apache.logging.log4j.Logger;
 import org.apache.velocity.Template;
 import org.apache.velocity.VelocityContext;
@@ -26,15 +71,13 @@ import org.sagebionetworks.template.repo.beanstalk.SecretBuilder;
 import org.sagebionetworks.template.repo.beanstalk.SourceBundle;
 import org.sagebionetworks.template.repo.beanstalk.image.encrypt.ElasticBeanstalkDefaultAMIEncrypter;
 import org.sagebionetworks.template.repo.beanstalk.image.encrypt.ElasticBeanstalkEncryptedPlatformInfo;
+import org.sagebionetworks.template.repo.cloudwatchlogs.CloudwatchLogsVelocityContextProvider;
 
 import com.amazonaws.services.cloudformation.model.Output;
 import com.amazonaws.services.cloudformation.model.Parameter;
 import com.amazonaws.services.cloudformation.model.Stack;
+import com.amazonaws.services.cloudformation.model.Tag;
 import com.google.inject.Inject;
-import org.sagebionetworks.template.repo.cloudwatchlogs.CloudwatchLogsVelocityContextProvider;
-import org.sagebionetworks.template.repo.cloudwatchlogs.LogDescriptor;
-
-import static org.sagebionetworks.template.Constants.*;
 
 public class RepositoryTemplateBuilderImpl implements RepositoryTemplateBuilder {
 
@@ -48,7 +91,6 @@ public class RepositoryTemplateBuilderImpl implements RepositoryTemplateBuilder 
 	Set<VelocityContextProvider> contextProviders;
 	ElasticBeanstalkDefaultAMIEncrypter elasticBeanstalkDefaultAMIEncrypter;
 	StackTagsProvider stackTagsProvider;
-	S3BucketBuilder bucketBuilder;
 	CloudwatchLogsVelocityContextProvider cwlContextProvider;
 
 	@Inject
@@ -56,8 +98,7 @@ public class RepositoryTemplateBuilderImpl implements RepositoryTemplateBuilder 
 										 RepoConfiguration configuration, LoggerFactory loggerFactory, ArtifactCopy artifactCopy,
 										 SecretBuilder secretBuilder, WebACLBuilder aclBuilder, Set<VelocityContextProvider> contextProviders,
 										 ElasticBeanstalkDefaultAMIEncrypter elasticBeanstalkDefaultAMIEncrypter,
-										 StackTagsProvider stackTagsProvider, S3BucketBuilder bucketBuilder,
-										 CloudwatchLogsVelocityContextProvider cloudwatchLogsVelocityContextProvider) {
+										 StackTagsProvider stackTagsProvider, CloudwatchLogsVelocityContextProvider cloudwatchLogsVelocityContextProvider) {
 		super();
 		this.cloudFormationClient = cloudFormationClient;
 		this.velocityEngine = velocityEngine;
@@ -69,14 +110,11 @@ public class RepositoryTemplateBuilderImpl implements RepositoryTemplateBuilder 
 		this.contextProviders = contextProviders;
 		this.elasticBeanstalkDefaultAMIEncrypter = elasticBeanstalkDefaultAMIEncrypter;
 		this.stackTagsProvider = stackTagsProvider;
-		this.bucketBuilder = bucketBuilder;
 		this.cwlContextProvider = cloudwatchLogsVelocityContextProvider;
 	}
 
 	@Override
 	public void buildAndDeploy() throws InterruptedException {
-		// build all of the buckets
-		this.bucketBuilder.buildAllBuckets();
 		
 		// Create the context from the input
 		VelocityContext context = createSharedContext();
