+++ src/main/java/org/sagebionetworks/template/s3/S3BucketBuilder.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.template.repo;
+package org.sagebionetworks.template.s3;
 
 public interface S3BucketBuilder {
