+++ src/test/java/org/sagebionetworks/template/repo/cloudwatchlogs/CloudwatchLogsConfigTest.java
@@ -1,14 +1,12 @@
 package org.sagebionetworks.template.repo.cloudwatchlogs;
 
-import com.google.inject.Guice;
-import com.google.inject.Injector;
+import static org.junit.Assert.assertNotNull;
+
 import org.junit.Test;
 import org.sagebionetworks.template.TemplateGuiceModule;
-import org.sagebionetworks.template.repo.beanstalk.EnvironmentType;
 
-import java.util.List;
-
-import static org.junit.Assert.*;
+import com.google.inject.Guice;
+import com.google.inject.Injector;
 
 public class CloudwatchLogsConfigTest {
