+++ src/test/java/org/sagebionetworks/template/repo/cloudwatchlogs/CloudwatchLogsConfigValidatorTest.java
@@ -1,21 +1,18 @@
 package org.sagebionetworks.template.repo.cloudwatchlogs;
 
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
-import org.sagebionetworks.template.repo.beanstalk.EnvironmentType;
+import static org.mockito.Mockito.when;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static org.junit.Assert.*;
-import static org.mockito.Mockito.when;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
+import org.sagebionetworks.template.repo.beanstalk.EnvironmentType;
 
 @RunWith(MockitoJUnitRunner.class)
 public class CloudwatchLogsConfigValidatorTest {
