+++ src/test/java/org/sagebionetworks/template/s3/S3BucketBuilderImplTest.java
@@ -1,12 +1,14 @@
-package org.sagebionetworks.template.repo;
+package org.sagebionetworks.template.s3;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.fail;
 import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.anyString;
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.when;
 import static org.sagebionetworks.template.Constants.PROPERTY_KEY_S3_BUCKETS_CSV;
 import static org.sagebionetworks.template.Constants.PROPERTY_KEY_STACK;
